refactoring (only 1 global var left!)
